BackEnd/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic[email]>=2.6.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
email-validator>=2.1.0
pydantic-settings>=2.0.0
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
PyJWT
